Undo base entity
